rename ggate from gate
