+++ ggate/README.md
@@ -1,0 +1,3 @@
+# ggate
+
+This is gRPC Gateway.
