+++ ggate/client/auth.go
