+++ ggate/client/image.go
