+++ ggate/data/group.go
@@ -2,15 +2,15 @@ package data
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/yoneyan/vm_mgr/gate/client"
+	"github.com/yoneyan/vm_mgr/ggate/client"
 	"log"
 )
 
-func GetNode(c *gin.Context) {
+func GetGroup(c *gin.Context) {
 	log.Println("------GetGroup------")
 
 	token := GetToken(c.Request.Header.Get("Authorization"))
-	result := client.GetNode(token)
+	result := client.GetGroup(token)
 
 	c.JSON(200, result)
