+++ ggate/data/node.go
@@ -2,15 +2,15 @@ package data
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/yoneyan/vm_mgr/gate/client"
+	"github.com/yoneyan/vm_mgr/ggate/client"
 	"log"
 )
 
-func GetGroup(c *gin.Context) {
+func GetNode(c *gin.Context) {
 	log.Println("------GetGroup------")
 
 	token := GetToken(c.Request.Header.Get("Authorization"))
-	result := client.GetGroup(token)
+	result := client.GetNode(token)
 
 	c.JSON(200, result)
